requirements.txt
alembic==1.13.0
annotated-types==0.6.0
APScheduler==3.10.4
bcrypt==4.1.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
garminconnect==0.2.12
garth==0.4.42
greenlet==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.4
Mako==1.3.0
MarkupSafe==2.1.3
oauthlib==3.2.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
withings-sync==4.1.0
WTForms==3.1.1
